requirements.txt
fastapi
uvicorn[standard]
httpx
python-telegram-bot>=21.0
